requirements.txt
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.2
django-admin==2.0.2
django-excel-response2==3.0.5
django-six==1.0.5
excel-base==1.0.4
isoweek==1.3.3
PyJWT==2.7.0
python-dateutil==2.8.2
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.6.3
tzlocal==5.0.1
urlopen==1.0.0
xlwt==1.3.0
